+++ client/src/components/FormInputRenderer.js
@@ -9,8 +9,18 @@ const StyledFormControl = styled(FormControl)`
   margin-bottom: 15px;
 `;
 
+// Renders a set of form input and calls a callback when one changes
+// Inputs:
+//    questions -> a list of question objects like:
+//      [{question: "This is question text?", type: "select", options: Array(2)}, ...]
+//    onValueChange -> a callback for when the value of an input changes, signature:
+//      onValueChange(questionId, questionData)
+//    values -> a map of values for the controlled form inputs like:
+//      { <questionId>: {response: <some value>, ... }, ...}
 const FormInputRenderer = (props) => {
   const { questions, onValueChange, values } = props;
+  console.log('>>>', questions, values);
+
   return (
     questions.map((question, idx) => {
       const questionId = idFromString(question.question);